move Metrics to utils
