+++ node/core/av-store/Cargo.toml
@@ -5,16 +5,19 @@ authors = ["Parity Technologies <admin@parity.io>"]
 edition = "2018"
 
 [dependencies]
+derive_more = "0.99.9"
 futures = "0.3.5"
-polkadot-subsystem = { package = "polkadot-node-subsystem", path = "../../subsystem" }
-polkadot-overseer = { path = "../../overseer" }
-polkadot-primitives = { path = "../../../primitives" }
-erasure = { package = "polkadot-erasure-coding", path = "../../../erasure-coding" }
+log = "0.4.8"
+
 kvdb = "0.7.0"
 kvdb-rocksdb = "0.9.1"
 codec = { package = "parity-scale-codec", version = "1.3.1", features = ["derive"] }
-log = "0.4.8"
-derive_more = "0.99.9"
+
+erasure = { package = "polkadot-erasure-coding", path = "../../../erasure-coding" }
+polkadot-overseer = { path = "../../overseer" }
+polkadot-primitives = { path = "../../../primitives" }
+polkadot-subsystem = { package = "polkadot-node-subsystem", path = "../../subsystem" }
+polkadot-node-subsystem-util = { path = "../../subsystem-util" }
 
 [dev-dependencies]
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "master" }
