+++ node/core/chain-api/Cargo.toml
@@ -9,6 +9,7 @@ futures = { version = "0.3.5" }
 sp-blockchain = { git = "https://github.com/paritytech/substrate", branch = "master" }
 polkadot-primitives = { path = "../../../primitives" }
 polkadot-subsystem = { package = "polkadot-node-subsystem", path = "../../subsystem" }
+polkadot-node-subsystem-util = { path = "../../subsystem-util" }
 
 [dev-dependencies]
 futures = { version = "0.3.5", features = ["thread-pool"] }
