+++ node/subsystem-util/Cargo.toml
@@ -10,19 +10,22 @@ async-trait = "0.1"
 derive_more = "0.99.9"
 futures = "0.3.5"
 futures-timer = "3.0.2"
-keystore = { package = "sc-keystore", git = "https://github.com/paritytech/substrate", branch = "master" }
 log = "0.4.8"
 parity-scale-codec = "1.3.4"
 parking_lot = { version = "0.10.0", optional = true }
 pin-project = "0.4.22"
+smallvec = "1.4.1"
+streamunordered = "0.5.1"
+
 polkadot-node-primitives = { path = "../primitives" }
 polkadot-node-subsystem = { path = "../subsystem" }
 polkadot-primitives = { path = "../../primitives" }
 polkadot-statement-table = { path = "../../statement-table" }
+
+sc-keystore = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sc-network = { git = "https://github.com/paritytech/substrate", branch = "master" }
-smallvec = "1.4.1"
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "master" }
-streamunordered = "0.5.1"
+substrate-prometheus-endpoint = { git = "https://github.com/paritytech/substrate", branch = "master" }
 
 [dev-dependencies]
 assert_matches = "1.3.0"
